+++ CHANGELOG.md
@@ -18,7 +18,8 @@
 - Added test cases for the saveGraphJSON function in `Controllers/Graph`
 - Added test cases for the getGraphJSON function in `Controllers/Graph`
 - Fix unused variable from `Graph.js`, formatting in `Container.js` and `GraphDropdown.js`, and eslint config
-- Refactor `graphImageResponse` in `Controllers/Graph` to use temporary files
+- Refactor `Controllers/Graph.hs` and `Controllers/Timetable.hs` to use temporary files
+- Remove unused `getTimetableImage` function in `Export/GetImages.hs`
 
 ## [0.7.2] - 2025-12-10